Front bug fixes and finalization

+++ src/Frontend/Delete.java
@@ -11,23 +11,26 @@ import connecthub.mappers.ContentMapper;
 import connecthub.mappers.ProfileMapper;
 import connecthub.mappers.UserMapper;
 import java.util.List;
+import javax.swing.JOptionPane;
 
 /**
  *
  * @author Mahinour Mohamed
  */
 public class Delete extends javax.swing.JFrame {
+
     User u;
     Profile p;
     Login l;
+
     /**
      * Creates new form Delete
      */
-    public Delete(User u,Profile p,Login l) {
+    public Delete(User u, Profile p, Login l) {
         initComponents();
-        this.u=u;
-        this.p=p;
-        this.l=l;
+        this.u = u;
+        this.p = p;
+        this.l = l;
     }
 
     /**
@@ -85,24 +88,37 @@ public class Delete extends javax.swing.JFrame {
     }// </editor-fold>//GEN-END:initComponents
 
     private void deleteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_deleteActionPerformed
-       UserMapper.delete(u.getID());
-       ProfileMapper.delete(u.getID());
-        List<Content>cont=ContentMapper.getAll();
-        for (int i = 0; i < cont.size(); i++) {
-            if(cont.get(i).getAuthorId()==u.getID())
-                ContentMapper.delete(cont.get(i).getID());
+        if (u == null || p == null || l == null) {
+            JOptionPane.showMessageDialog(this,
+                    "User or Newsfeed data is missing. Please log in again.",
+                    "Error",
+                    JOptionPane.ERROR_MESSAGE);
+            return;
         }
-       FrontProfile f= FrontProfile.getInstanceOf();
-      f.setVisible(true);
-      f.setLocation(null);
-      setVisible(false);
-        
+        try {
+
+            UserMapper.delete(u.getID());
+            ProfileMapper.delete(u.getID());
+            List<Content> cont = ContentMapper.getAll();
+            for (int i = 0; i < cont.size(); i++) {
+                if (cont.get(i).getAuthorId() == u.getID()) {
+                    ContentMapper.delete(cont.get(i).getID());
+                }
+            }
+            FrontProfile f = FrontProfile.getInstanceOf();
+            f.setVisible(true);
+            f.setLocation(null);
+            setVisible(false);
+        } catch (Exception e) {
+
+        }
+
     }//GEN-LAST:event_deleteActionPerformed
 
     /**
      * @param args the command line arguments
      */
-    
+
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton delete;
     private javax.swing.JLabel jLabel1;
